requirements.txt
asgiref==3.7.2
dj-rest-auth==5.0.1
Django==5.0a1
djangorestframework==3.14.0
pytz==2023.3.post1
sqlparse==0.4.4
